+++ src/components/topbar.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import TopNavigation from '@cloudscape-design/components/top-navigation';
-import TickerSearchbar from './ticker-searchbar';
+import TickerSearchbar from './ticker-searchbar/ticker-searchbar';
 
 function Topbar() {
   return (
